+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Add new `productImageTag--zoom` and `productImageTag--main` CSS Handles to Product Images component.
+
 ## [3.121.0] - 2020-07-27
 ### Added
 - Expose `SanitizedHTML` component.